+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 1. [](#bugfix)
     * Fixed extra space in Flex admin pages
     * Fixed folder creation with parent other than root [#66](https://github.com/trilbymedia/grav-plugin-flex-objects/issues/66)
+    * Fixed task redirects in sub-folder multi-site environments
 
 # v1.0.0-rc.15
 ## 07/22/2020